set a subsurface buffer to ec even if parent is invisible.
We want to show a newly-attached subsurface buffer when parent becomes
visble. However, if we keep a subsurface buffer to cache when parent is
invisible, it makes the issue that old buffer can be shown on screen when
parent becomes visible. So we have to set a subsurface buffer to e_pixmap
even if parent is invisble. Instead, we just hide ec->frame until parent's
frame is visible.

Change-Id: I1ae07dba21125b5544ac4ef55ba36e9d85ef1bab