wl_desktop_shell: E_Xdg_Shell to support ping-pong for the xdg_shell
ping-pong protocol of the xdg_shell is working on the wl_client
and not on the wl_surface. However, ping-pong of the enlightenment
is on the wl_surface (E_Client). E_Xdg_Shell is designed for connecting
them. It has a list of E_Client which sent ping event, and if the
client is in responsive, then it sets ping_ok for all E_Clients.

And unnecessary shell_resources and xdg_shell_resources hash are removed.

Change-Id: I0e1e1815377f5488afeb3c1ccff0d8ba441180e1