e_comp_object: do not clear surface when buffer is destroied
enlightenment has the right to deal with a buffer even if a client requested
wl_buffer@destroy. So the enlightenment doesn't need to clear surface or
to delete evas object for wl_buffer@destroy when the enlightenment wants to
continue to display the buffer, enlightenment can use the destroied buffer
until there are no task to need the destroied buffer.

Change-Id: I84845bbb4a828ac763142386e65361b4e599598d
Signed-off-by: MinJeong Kim <minjjj.kim@samsung.com>