e_comp_object: fixed size update problem for H/W composited cw.
The size of non-compositing window can be changed, so there is a
need to check that cw is H/W composited if cw is not redirected.
And of course we have to change size of evas object of H/W composited cw,
otherwise cw can't receive input events even if it is shown on the screen.

Change-Id: I9a126bdcb7c1ad0be11e9fd3c4e8df3804f4a1fc