e_comp_wl: use the transparent color for below_obj
It's more reasonable to use the transparent color instead of black because
we can show the alpha color of the 24 depth topmost window. Even if we use
the transparent color, we still can show the black rectangle behind the
topmost window because the transparent color will be copied to canvas.

Change-Id: I6c00430e03415be49b10e803ac31cb63950a94cd