[e_info] use the uint64_t type instead the int32_t for window id
We have some error when a window id exceeds INT_MAX, because
the Ecore_Window is the uintptr_t type and the  window id the
int32_t type. I think we need use the uint64_t(for arm7l and arm64)
type instead the int32_t type for window id.

Change-Id: I534965bf2d0e50fb23adbf266224c335c08b7a60
Signed-off-by: Konstantin Drabeniuk <k.drabeniuk@samsung.com>