visibility: add E_VIS_ICONIFY_STATE_UNKNOWN enumeration
There was a bug that the window is not shown althogh it requested to activate.
It was because E_Client iconic status mismatch with E_Vis_Iconify_State
due to wrong initialize of E_Vis_Client with E_VIS_ICONIFY_STATE_UNICONIC.
so we add a E_VIS_ICONIFY_STATE_UNKNOWN and set the initail state to this.

Change-Id: I4d640b5f761042ffb70de4a7227ae4cf5787c4aa