e_comp_object: reset and move the position of an object
Even cw->x and cw->y are already set to -99999 value, if new value is same
with old value, cw->x and cw->y can't be changed by evas_object_move().
So only when the new position is different with old value, we reset and
move the position of an object.

Change-Id: I0dbab086cae9013ccbcb1e0595ba0d8aa61f1d5b