e_comp_wl: adding window rotation information to surface's buffer transform
To show a client buffer via a HW overlay, the client buffer status is exactly
same with WR 90(Window Rotation 90) and SR 90(Screen Rotation). So it's
reasonable that the surface's buffer transform contains the information how
a client draws objects on an attached buffer regarding window rotation.

Then, we have to distinguish the window rotation and screen rotation information
from the surface's buffer transform in server side.

Change-Id: I34128a6c192fcb4fb8527f9f1fb1984a5950bc4d