deiconify_approve: check transient child while finding above window when the win is lowered/iconified
When an app requested to lower/iconify window, we check its above window.
If above window was full size non-alpha window, then we didn't operate the deiconify_approve functionality.
But there was no code to check whether the above window was child of the window or not.
If the above window is child window, then we skip it.

Change-Id: I9f78b8e453ba1a937ca650ba02344c9c6327d45c