remove the link of wl_resource destory listener
wl_resource was use the wl_priv_signal for destroy signal from new version of wayland.
unlike the wl_signal, the link of signal should be removed in the destory callback function.
If not, it could the cause of memory crash when instance of link was freed without unlink
in the destroy callback.

Change-Id: Id98e0c6262adbf6c9b3df6f272ab11a199cefb4c