e_policy_visibility: change vc->state to UNICONIC before calling _e_vis_job_eval
There was a bug that the window which rendered already was added an uniconify_render job list again
because its vc->state was RUNNING_UNICONIFY.
For this, the log "VISIBILITY | TIMEOUT ~~~" was printed sometimes.
So, we changed its vd->state to UNICONIC to prevent adding an uniconify_render job list.

Change-Id: Iae33954ef675bd62fee788e4501724049f0aa219