e_policy_visibility: fix bug for deiconify_approve transient child
There was bug for handing deiconify approve when tree windows set transient_for another windows.
For example, A win set transient to B win, B win set transient to C win. (C is parent of B, and
B is parent of A.)
In this case, if C win requests to activate, C, B, A should be run deiconify_approve and
it should be worked well. However A, B always take place timeout for deiconify_approve.
This patch resolve this bug.

Change-Id: I487068a32ebbf51f7221c4a1eacf549e7fdc088f