e_config: added support for evas_norender to avoid memory leaks for evas objects
The memory leak in the evas could occur if E doesn't perform evas rendering.
It is because cleanup for evas object is processed in the rendering stage of
evas. Thus if E is always performing HWC for compositing the new evas object,
then leak can occurs on the evas side even after deletion of that object.

In order to resolve it, we have added comp_canvas_norender configuration value.
If user enables this configuration, then E always performs evas_norender in idle
to cleanup evas objects which had not been rendered through evas rendering.

Change-Id: I89dd8d931b75f78ca2114e493ca7b46eb81bf28e