[hwc_opt_2] implement smooth transition from device to client
If a window with the device composition type changes a type
to the client composition type (by the e20 demands or by
the driver demands) we got the flickering as a buffer owned by
this window can't be composited to the fb_target immediately.

We start to composite a buffer owned by an e_window which
composition type has been tried to change from device/video
to client and after we got it composited we inform hwc-extension
about this to let it unset hw overlay owned by this
e_window(hwc_window).

Change-Id: I85be90f3c8af5a2b6166920f9a048bf671ef8694
Signed-off-by: Sergey Sizonov <s.sizonov@samsung.com>