e_comp_object: fixed a problem that final clip region of the mouse cursor object could be wrong after rotating
Due to an initial position of cw->clip that is (0, 0), it could occur that the
top of the mouse cursor object could be clipped after rotating. Ideally, this
issue should be fixed at the evas rendering engine side, but extending the position
to (-9999,-9999) from (0, 0) is enough for resolving this issue.

Change-Id: Icdf1bfbb428945827b542be1475fac2b32de28bb