fix bug for sending visibility event
There was a bug that the window which was waiting for rendering got the visibility
event sometimes as below.
"pre-unobscured -> fully-obscured -> unobscured"
For this, application's state changed to resume, pause, resume.
So, we fixed code not to send fully-obscured event when it is waiting for rendering.

However, the fully-obscured event is sent when the lcd is turned off.

Change-Id: Ifc400dfc3314407d6fe54b83f64cd40abcfb7eee