return if there are not updates and there is displaying surface
If there are not updates and there is not displaying surface it means
the hwc_window's composition type just became TDM_COMPOSITION_DEVICE.
Buffer for this hwc_window was set in the previous e_hwc_window_fetch
but ref for buffer didn't happen because hwc_window had
TDM_COMPOSITION_CLIENT type. So e20 needs to make ref for the current
buffer which is set on the hwc_window.

Change-Id: Iadbf884135c8d8bfb71e29141b169c4c984e1e6b
Signed-off-by: Konstantin Drabeniuk <k.drabeniuk@samsung.com>