e_plane: set unset_counter to 1 when fb_target is unset to hwc
unset_counter is set to 0 when fb_target used by hwc because
compositor is disabled and then plane doesn't need to sync with fb_target.
but unset_candidate plane should sync with fb_target when fb_target is unset to hwc.
so if plane is unset_candidate and unset_counter is 0, set unset_counter to 1
for sync with compositor

Change-Id: I62bae00c12fd3d0e68c9986b2a8b0d8f3ebec952