add avoid_afill option
The avoid_afill option is for initializing ecore_evas engine.

When the avoid_afill is 1, e20 uses the gl_tbm_ES engine.
The gl_tbm_ES engine does not fill 0xff the alpha channel even if
the depth of the ecore_evas is 24 depth and the bpp of the buffer
is 32bits.
Otherwise, the gl_tbm engine which is used when the avoid_afill is 0
(default value) fills the 0xff at the same condition.

p.s. this patch is temporary. therefore it will be revoved later.

Change-Id: I8af6e3be883d800687a8cfd893bd5f5c2555eddf