e_policy_visibility: fix bug for sending pre-unobsucred event
There was a bug that the pre-unobscured event was sent to the window which was
obscured by alpha-opaque window.

[top] alpha-opaque window -> A win -> B win [bottom]

In above stack, when "A win" is lowered/hidden, then "B win" is chagned to uniconic
and must maintain its visibility to fully-obscured.
So, e must not send a pre-unobscured event to "B win".

But before this patch, e sent a pre-unobscured event to "B win".
We fixes this bug.

Change-Id: If6835790442d619ec0aac2dfa1659e5f0f96729a