e_comp: job handler(_cb_update) make no update if ec pixmap size is wrong
if commit wl_buffer is null buffer attached(means explicit hide request)
or if commit wl_buffer is invalid,
than cw object make no updates and let it as it is.
if server make cw object unredirected in above condition, as a result
it never updated on screen on show request, due to unredirect condition.

Change-Id: I4003d360982ed5e821012ff5bd6ba2ac72ba875e