e_policy_visibility: add a E_VIS_ICONIFY_STATE_RUNNING_UNICONIFY_RENDER_DONE state
There was a bug that the window which was pending uniconify by visibility didn't raise
even if it requested to activate.

[TOP] A win -> B win(32bit) -> C win [BOTTOM]

Above window stack,
when "A win" request to lower, then "B win" and "C win" are running uniconify_pending with
not_raise flag.
After "C win" is rendered, its internal state changed to E_VIS_ICONIFY_STATE_UNICONIC.
But e_client_uniconify() is not called because e is waiting for rendering of "B win".

In this time, if "C win" requests to activate, then this request is not added to pending list
because the state of "C win" is E_VIS_ICONIFY_STATE_UNICONIC.

As a result of this, the window stack is changed as below.
[TOP] B win(32bit) -> C win -> A win [BOTTOM]

But, the expected stack is as below.
[TOP] C win -> B win(32bit) -> A win [BOTTOM]

To fix this problem, we add a additional state - E_VIS_ICONIFY_STATE_RUNNING_UNICONIFY_RENDER_DONE.
It means that the window is rendered, but it remains iconic state.
So, the activate request of "C win" is added to the pending list in above case.

The E_VIS_ICONIFY_STATE_RUNNING_UNICONIFY_RENDER_DONE state
is changed to E_VIS_ICONIFY_STATE_UNICONIC after all pending windows are rendered.

Change-Id: I3fff995dbbad7f438e952616bc6840e7a92e2424