e_policy_visibility: check map state of above window in _e_policy_check_above_alpha_opaque
There was a bug that a window which was obscured by alpha opaque window was iconified.

[TOP] A win (32bit) -> B win (unmapped) -> C win [BOTTOM]
If "A win" change its opaque state to true, we expect that "C win" maintain unicoic state
and just change visibility value to fully-obscured.

But, there was no check code to above window's map state, "C win" was changed to iconic
and fully-obscured.

This patch fixes this problem.

Change-Id: I450521110681929ddf7898eea30db685e7a8609a