e_xdg_shell_v6: change condition for client's surface commit
whenever client set and unset its user geometry,
configure event with width and height(zero or screen size) is sent to client.
when user geometry gets unset, sometimes client commit serveral times
with invalid size buffer, and right size buffer after a while eventually.
in that case e20 was not adjust its ec->frame size refer to delayed commit(the right size)
because of this issue, make client to be changed by their commit buffer size

in addition to that, added new condition for error
(a commit which attempts configure prior to the first xdg_surface.configure is error)

Change-Id: I1e7279db49fa1fb7bce0be63c82f143b8394c6ca