e_client: add a way to prevent generating event duplicated
There was a bug that events (such as E_EVENT_CLIENT_ADD) were sent duplicated in some cases.
So, we added a flag to e_client structure to check whether an event was generated already.
For example, if a E_EVENT_CLIENT_ADD was generated and sent, then it could not be send until
generating E_EVENT_CLIENT_REMOVE event.

For this, we could pair sending ADD/REMOVE, SHOW/HIDE events.

Change-Id: Id42d9c0a6594d0abef270a4970745fa8f23ebf1e