e_policy: fix bug for handling iconify/uniconify fully-obscured window
There was a bug that the fully-obscured window was not iconified in visibility hook.
If a skip_iconify window is placed above on the fully-obsucred window, then the fully-obscured
window could not be iconified because the skip_iconify window is not considered to find above window.

So, If below case, "win B" is not iconified and could get the focus wrong.

[top]
Win A (24 bit skip_iconify)
Win B (24 bit fully-obscured, but not iconic)
[bottom]

Change-Id: Ibd72f426d68661b9f767914db6f64915959adaad