e_policy_visibility: added new hook points and check grab for lower,hide job
- E_POL_VIS_HOOK_UNICONIFY_RENDER_RUNNING hook point is newly added. This
can inform someone that, enlightenment is waiting for render process
in order to uniconify an E_Client. And it means 'someone' who subscribes this
hook point is able to know uniconify process of the E_Client is in pending state.

- E_POL_VIS_HOOK_LOWER hook point is newly added. This hook point provides
a chance that someone can grab 'lower' job.

- E_POL_VIS_HOOK_HIDE hook point is newly added. This hook point provides
a chance that someone can grab 'hide' job.

- additional check for grab on lower API.
- additional check for grab on hide interceptor callback.

Change-Id: I6a928e782fbdca25c789dd25efd568d51afbe37e
Signed-off-by: MinJeong Kim <minjjj.kim@samsung.com>