e_serivce_launcher : support launcher service
e_serivce_launcher supports a launcher application who proccesses launching
animation. The launcher application should be privileged to be assigned
a role of launcher service and if the launcher is passed the privilege test,
enlightenment would accepts connection of tws_launcher_service from the launcher.
The launcher application and enlightenment are able to control launching
animation with those connection.

Change-Id: I86366ee188dc8f431a2405f4f2c2f3b6fae308ab
Signed-off-by: MinJeong Kim <minjjj.kim@samsung.com>