e_client: fix bug for setting focus to invalid window in visibility calculation
When the skip-focus set and full size window was shown on the higher layer,
there was a bug that the normal window got and lost focus when it requested to activate.
It had not to get the focus because it was obscured by the higher layer window.
For this bug, the normal window got both focus-in and focus-out event.

To resolve this problem,
we added checking code whether the top visible full size window set skip-focus or not.

Change-Id: Iaa2c7564187f39e260bf33ab315ef4b02e4431df