use e_client_layer_set api for client stack change
if it has only purpose for changing the evas obj layer(such as ibar
effect animation temporarily, or quickpanel mover generation)
use evas_object_layer_set, but if it is about the client layer policy change
use e_client_layer_set api. because it handles the exception of
layer_pending and layer_block.

Change-Id: Ia73dba859e65e52057d526ef78e3727c2304ba9d